Handle values with var() when parsing font for Canvas/FontFaceSet
When setting the 'font' attribute, it must parse according to the
grammar of <font> (minus CSS-wide keywords) [1]. Values with var()
does not match the grammar of <font>, but the parsing function we
use allow it anyway, since var()-containing values have special
handling during regular CSS declaration parsing.

There are multiple call sites that parse a <font> and then try to
fix up the result (with IsCSSWideKeyword). This CL adds
CSSParser::ParseFont which does the check for CSS-wide keywords, and
now also for the pending-substitution value (which is what you get
if you parse a var()-containing sequence as a shorthand).

This fixes DCHECKs/undefined behavior when using var() in a canvas
font attribute, or in FontFaceSet.load().

Bug: 1131922

[1] https://html.spec.whatwg.org/multipage/canvas.html#dom-context-2d-font

Change-Id: Ic906932fa14ad46683cf97edc78e2929020e0fd5